feat(styles): moved css to assets folder
