+++ src/AutoVizuA11y.tsx
@@ -16,7 +16,7 @@ import { generateDescriptions } from "./components/descriptions/DescriptionsGene
 import { insightsCalculator } from "./utils/insightsCalculator";
 import { descriptionsChanger } from "./components/descriptions/Descriptions";
 
-import "./style/AutoVizuA11y.css";
+import "./assets/style/AutoVizuA11y.css";
 
 type AutoDescriptionsProps = {
 	dynamicDescriptions?: boolean;
